Mongo DB implementation

+++ src/main/java/com/cmad4/team8/blogrestcontroller/post/api/Comments_Interface.java
@@ -7,9 +7,9 @@ import com.cmad4.team8.blogrestcontroller.exceptions.CommentGeneralException;
 public interface Comments_Interface {
 	
 	public void Create(comments c) throws InvalidCommentException, CommentGeneralException;
-	public comments read(int c_id) throws CommentNotFoundException, CommentGeneralException;
-	public List<comments> readallbypost(int p_id, int pagenum) throws CommentNotFoundException, CommentGeneralException;
+	public comments read(Long c_id) throws CommentNotFoundException, CommentGeneralException;
+	public List<comments> readallbypost(Long p_id, int pagenum) throws CommentNotFoundException, CommentGeneralException;
 	public List<comments> readallbyuser(String login_id, int pagenum) throws CommentNotFoundException, CommentGeneralException;
-	public void removeCmt(int c_id) throws CommentNotFoundException, CommentGeneralException;
+	public void removeCmt(Long c_id) throws CommentNotFoundException, CommentGeneralException;
 
 }
